Update patch set 1
Patch Set 1:

(9 comments)

As stated, it is not easy to review partial code. So i am not able to understand the model changes if i not see the changes in the dependent code fragments. Commits should be closed compilable functional blocks of a ticket.

Patch-set: 1
Reviewer: Gerrit User 1015828 <1015828@d5d70762-12d0-45a1-890d-524b12d3f735>