Refine display of legend entries.
Merging this commit will change the behaviour of the legend to be sorted by the index as the rules appear in the highliting rules table.
For rules that have both a Station and a Delivery rule, the lower of the two indecies is used.
Jumping can now only occur if a user deliberately sorts the delivery and station highliting rules differently.

Additionally the graphviewcomponent and highliting service functions that needed to be ajusted for this change where refactored.

Ticket: #785