Zoom To Rectangle #839
Adds functionality for users to quickly zoom to a rectangular region in the graph by spanning a rectangular frame (similar to the selection of graph elements) while pressing the key 'Strg' / Cmd'.
This zooms to the area of the rectangle.

If the side panel is open, then the zoomed area is fit to the visible area only.

Ticket: #545

* WIP first ideas for how to approach this.
* Add rect zoom functionallity.
* WIP: I am not certain wether I am getting absolute positions or a relative offset. Might revert this later.
* Different zoom function.
* added unzoomed to zoomed position handling
* Add fitToVisible area with default true.