Fixes upload model bug in safari (#895) and adds UTX upload button
* Fixes upload model bug in safari

In Safari, users were only able to upload data when no example or other data had been loaded before.This was due to the input type=file being programmatically triggered in an async function, which is a security violation for safari. This moves the check the async function performs onto the upload menu opening button, so that we don't lose the trusted context for the input type=file click.

Ticket: #761

* Adds UTX upload button (#896)

The button is currently disabled and a tooltip informs the user that this feature will soon be available. Please remove tooltip and enable button once we've activated the UTX upload.

Ticket #887