Ticket #888: Adds data upload warnings symbol and modal - Part 1 (#898)
Ticket #888: Adds data upload warnings symbol and modal

The toolbar component can now receive props stating if a data upload resulted in warnings. If this prop is true, a warning symbol will be displayed next to the file name in the toolbar. When a user clicks on the symbol, a modal opens. The modal is currently empty but shall eventually display a list of warnings.

Ticket #888