refines treatment of utx schema rules on utx import
The type constraints for Date, DateTime, Duration1, Duration2, Email and
Link are now treated as soft contraints. Format violations of these types
are tried to be fixed but not deleted otherwise.
Violations of fixes are reported to the user.

The positive amount contraint was removed from the schema, since it is not
part of UTX_TABLES_20240208

ticket: #913