fixes missing map (after window resize & hidden container) bug
Bug cause: Insufficient window resize handling in geomap component

Details: The internal map size update in openlayer map (triggered by window resize)
does not take into account that the container might be invisible. The geomap component
handles only window resize events if the geomap is visible ==>
the internal map resize might set the map size to 0 (if the geo map is hidden), and the
geomap component does not correct the size after it becomes unhidden again

Solution:
If an window resize event occures and the geomap is hidden, a suspendedFlag is set.
If the geomap component is visible again, the map is resized (if the flag was set).

Other remarks:
- The resize sensor directive was replaced by the visibility sensor directive:
  The graph view does not need a size sensor at all because the cy graph updates its size automatically on
  window resize and is not affected by hidden containers

ticket: #989